Solution to Check if a Number is Prime or Not

+++ Check if a Number is Prime or Not@@6788dd63490637cfc44e1ba2/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main()
 {
-    int num,i,Jasoos;
-    scanf("%d",&num);
+    float num,i,Jasoos;
+    scanf("%f",&num);
     for(i=2;i<num;i++)
     {
        if(num%i==0)
